bigquery: begin splitting snippets out, starting with datasets.
Make BigQuery snippets more like style guide.

Also, adds a snippets/testutil for generating resource names.
Let me know if you'd prefer this to be in internal/testutil
or somewhere else.
