+++ bigquery/snippets/dataset/bigquery_list_datasets_by_label.go
@@ -1,0 +1,45 @@
+// Copyright 2019 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package dataset
+
+// [START bigquery_list_datasets_by_label]
+
+import (
+	"context"
+	"fmt"
+	"io"
+
+	"cloud.google.com/go/bigquery"
+	"google.golang.org/api/iterator"
+)
+
+func listDatasetsByLabel(client *bigquery.Client, w io.Writer) error {
+	ctx := context.Background()
+	it := client.Datasets(ctx)
+	it.Filter = "labels.color:green"
+	for {
+		dataset, err := it.Next()
+		if err == iterator.Done {
+			break
+		}
+		if err != nil {
+			return err
+		}
+		fmt.Fprintf(w, "dataset: %s\n", dataset.DatasetID)
+	}
+	return nil
+}
+
+// [END bigquery_list_datasets_by_label]
