+++ bigquery/snippets/dataset/bigquery_updata_dataset_access.go
@@ -1,0 +1,49 @@
+// Copyright 2019 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package dataset
+
+// [START bigquery_update_dataset_access]
+
+import (
+	"context"
+
+	"cloud.google.com/go/bigquery"
+)
+
+func updateDatasetAccessControl(client *bigquery.Client, datasetID string) error {
+	ctx := context.Background()
+	ds := client.Dataset(datasetID)
+	meta, err := ds.Metadata(ctx)
+	if err != nil {
+		return err
+	}
+	// Append a new access control entry to the existing access list.
+	update := bigquery.DatasetMetadataToUpdate{
+		Access: append(meta.Access, &bigquery.AccessEntry{
+			Role:       bigquery.ReaderRole,
+			EntityType: bigquery.UserEmailEntity,
+			Entity:     "sample.bigquery.dev@gmail.com"},
+		),
+	}
+
+	// Leverage the ETag for the update to assert there's been no modifications to the
+	// dataset since the metadata was originally read.
+	if _, err := ds.Update(ctx, update, meta.ETag); err != nil {
+		return err
+	}
+	return nil
+}
+
+// [END bigquery_update_dataset_access]
