+++ bigquery/snippets/dataset/bigquery_update_dataset_description.go
@@ -1,0 +1,41 @@
+// Copyright 2019 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package dataset
+
+// [START bigquery_update_dataset_description]
+
+import (
+	"context"
+
+	"cloud.google.com/go/bigquery"
+)
+
+func updateDatasetDescription(client *bigquery.Client, datasetID string) error {
+	ctx := context.Background()
+	ds := client.Dataset(datasetID)
+	meta, err := ds.Metadata(ctx)
+	if err != nil {
+		return err
+	}
+	update := bigquery.DatasetMetadataToUpdate{
+		Description: "Updated Description.",
+	}
+	if _, err = ds.Update(ctx, update, meta.ETag); err != nil {
+		return err
+	}
+	return nil
+}
+
+// [END bigquery_update_dataset_description]
